+++ kernel/frame.h
@@ -8,6 +8,7 @@
 #define VM_FRAME_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "instruction.h"
 
 #define WORK_STACK_SIZE_MAX 128
@@ -24,24 +25,24 @@ typedef struct frame frame_t;
 struct frame {
     type_frame_t type_frame;
 
-    long long index_frame;
+    int index_frame;
     char *name;
 
-    long long number_args;
+    int number_args;
 
     instruction_t **instructions;
-    long long number_instructions;
+    int number_instructions;
 
-    long long *work_stack;
-    long long index_first_element_work_stack;
+    int64_t *work_stack;
+    int index_first_element_work_stack;
     bool *is_work_stack_element_object;
 
-    long long *local_pool;
+    int64_t *local_pool;
     bool *is_local_pool_element_object;
 };
 
-frame_t *constructor_frame(long long index_frame, char *name, type_frame_t type_frame, instruction_t **instructions,
-                           long long number_instructions, long long number_args);
+frame_t *constructor_frame(int index_frame, char *name, type_frame_t type_frame, instruction_t **instructions,
+                           int number_instructions, int number_args);
 
 void destructor_frame(frame_t *frame);
